# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JARVIS",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/agents": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Agent",
        "operationId": "create_agent_v1_agents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_v1_agents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agent_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent",
        "operationId": "get_agent_v1_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "agents"
        ],
        "summary": "Update Agent",
        "operationId": "update_agent_v1_agents__agent_id__patch",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "summary": "Delete Agent",
        "operationId": "delete_agent_v1_agents__agent_id__delete",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agent_id}/versions/{version}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Version",
        "operationId": "get_agent_version_v1_agents__agent_id__versions__version__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentVersion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agent_id}/run": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Run Agent",
        "description": "Blocking run \u2014 enqueue, then wait for the worker's segment to end. The\nsubscribe replays anything the worker already wrote, so there is no race\nbetween enqueueing and listening. A pause (S10) ends the segment like a\nterminal: the route returns the awaiting_input row and the client\nresumes with POST /executions/{id}/resume.",
        "operationId": "run_agent_v1_agents__agent_id__run_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agent_id}/stream": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Stream Agent",
        "description": "SSE run: enqueue, then frame every event the worker writes. Resume with\n`Last-Event-ID` (durable cursor) plus `run_id` in the body \u2014 replay and\nlive tail are the same cursor space (PgEventStream tails the DB, so the\nstream survives this process). Ends with the segment's last event\n(terminal or pause, S10): a pause closes the stream and the client\nresumes with POST /executions/{id}/resume.",
        "operationId": "stream_agent_v1_agents__agent_id__stream_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentDetail": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/AgentDefinition"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "definition",
          "versions"
        ],
        "title": "AgentDetail"
      },
      "AgentList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentDefinition"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AgentList"
      },
...
      "AgentUpsertRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "system_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Prompt"
          },
          "user_prompt_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Prompt Template"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ToolBinding"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          },
          "strategy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StrategyConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "memory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_iterations": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 32.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Iterations"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "output_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Schema"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AgentUpsertRequest",
        "description": "Partial create/update payload \u2014 everything except name/model/strategy\nis optional and falls back to the AgentDefinition defaults."
      },
      "AgentVersion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "snapshot": {
            "$ref": "#/components/schemas/AgentDefinition"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "version",
          "snapshot"
        ],
        "title": "AgentVersion",
        "description": "Immutable, append-only snapshot of an AgentDefinition.\n\n`snapshot` is the replay source of truth for every execution that\nreferences this version (ADR 0002: snapshot-vs-reference)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemoryConfig": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "max_messages": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Messages",
            "default": 20
          },
          "session_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MemoryConfig"
      },
...
      "ModelRef": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "credential_ref": {
            "anyOf": [
              {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/EnvCredentialRef"
                  },
                  {
                    "$ref": "#/components/schemas/StoredCredentialRef"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "env": "#/components/schemas/EnvCredentialRef",
                    "stored": "#/components/schemas/StoredCredentialRef"
                  }
                }
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "model"
        ],
        "title": "ModelRef",
        "description": "Reference to a model. `credential_ref` is a *reference* (env-var name\nor stored credential id) \u2014 credential material never enters the domain,\na snapshot, or an API response (ADR 0006 \u00a71)."
      },
...
      "RunRequest": {
        "properties": {
          "input": {
            "type": "string",
            "minLength": 1,
            "title": "Input"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "variables": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variables"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input"
        ],
        "title": "RunRequest",
        "description": "Body for /run and /stream. `run_id` is stream-resume only: attach to\nan existing run's stream instead of starting a new one."
      },
      "RunResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "awaiting_input",
              "succeeded",
              "failed",
              "cancelled",
              "timed_out"
            ],
            "title": "Status"
          },
          "input": {
            "type": "string",
            "title": "Input",
            "default": ""
          },
          "agent_version_id": {
            "type": "string",
            "title": "Agent Version Id",
            "default": ""
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "final_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Message"
          },
          "total_usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "iterations": {
            "type": "integer",
            "title": "Iterations",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Kind"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "event_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Cursor"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "agent_id",
          "status"
        ],
        "title": "RunResult"
      },
...
      "StrategyConfig": {
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1,
            "title": "Type"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type"
        ],
        "title": "StrategyConfig"
      },
...
      "ToolBinding": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ToolBinding"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionSummary": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "id",
          "label",
          "created_at"
        ],
        "title": "VersionSummary"
      },
...
    }
  }
}